fix: move ConfigJupyterLabView to a folder
